Move CLI configuration to separate file.
