+++ lib/config.js
@@ -2,108 +2,10 @@
  * @authot tupadr3
  */
 const cliArgs = require('command-line-args'),
-	os = require('os'),
+	cliOptions = require('./cliOptions.js'),
 	path = require('path');
 
 function initConfig() {
-	let binRsvgDft = os.platform() === 'linux' ? 'rsvg-convert' : 'assets/bin/rsvg-convert.exe';
-
-	let cpuCount = os.cpus().length;
-	if (cpuCount > 2) {
-		cpuCount = Number((cpuCount / 2).toFixed(0));
-		if (cpuCount > 6) {
-			cpuCount--;
-			cpuCount--;
-		}
-	}
-
-	const cliOptions = [
-		{
-			name: 'limit',
-			type: Number,
-			alias: 'l',
-			defaultValue: 0
-		},
-		{
-			name: 'concurrency',
-			type: Number,
-			alias: 'c',
-			defaultValue: cpuCount
-		},
-		{
-			name: 'progress',
-			type: Boolean,
-			alias: 'p',
-			defaultValue: true
-		},
-		{
-			name: 'verbose',
-			type: Boolean,
-			alias: 'v',
-			defaultValue: false
-		},
-		{
-			name: 'formats',
-			type: String,
-			multiple: true,
-			defaultValue: ['png', 'svg', 'puml']
-		},
-		{
-			name: 'github',
-			type: Boolean,
-			defaultValue: false
-		},
-		{
-			name: 'colors',
-			type: String,
-			multiple: true,
-			defaultValue: ['black']
-		},
-		{
-			name: 'sizes',
-			type: Number,
-			multiple: true,
-			defaultValue: [48]
-		},
-		{
-			name: 'icons',
-			type: String,
-			multiple: true,
-			defaultValue: []
-		},
-		{
-			name: 'fonts',
-			type: String,
-			multiple: true,
-			defaultValue: []
-		},
-		{
-			name: 'binPlantuml',
-			type: String,
-			defaultValue: 'assets/bin/plantuml.jar'
-		},
-		{
-			name: 'binRsvg',
-			type: String,
-			defaultValue: binRsvgDft
-		},
-		{
-			name: 'dist',
-			type: String,
-			defaultValue: 'dist'
-		},
-		{
-			name: 'temp',
-			type: String,
-			defaultValue: '.tmp'
-		},
-		{
-			name: 'devel',
-			type: Boolean,
-			defaultValue: false
-		}
-	];
-
 	const cfg = cliArgs(cliOptions);
 	const fontsDef = require('./fonts').def();
 	let fonts = cfg.fonts;
